+++ src/idunno.Bluesky/Actor/SavedFeedPreferenceV2.cs
@@ -8,45 +8,46 @@ namespace idunno.Bluesky.Actor;
 /// <summary>
 /// Encapsulates an actor's feed preferences.
 /// </summary>
-public record SavedFeedPreferences2 : Preference
+public record SavedFeedPreferencesV2 : Preference
 {
     /// <summary>
-    /// Creates a new instance of <see cref="SavedFeedPreferences2"/>.
+    /// Creates a new instance of <see cref="SavedFeedPreferencesV2"/>.
     /// </summary>
     /// <param name="items">A list of feed preferences.</param>
     [JsonConstructor]
-    public SavedFeedPreferences2(IReadOnlyList<SavedFeedPreference2> items)
+    public SavedFeedPreferencesV2(IReadOnlyList<SavedFeedPreferenceV2> items)
     {
         if (items is null)
         {
-            Items = new List<SavedFeedPreference2>().AsReadOnly();
+            Items = new List<SavedFeedPreferenceV2>().AsReadOnly();
         }
         else
         {
-            Items = new List<SavedFeedPreference2>(items).AsReadOnly();
+            Items = new List<SavedFeedPreferenceV2>(items).AsReadOnly();
         }
     }
 
     /// <summary>
     /// Gets a readonly list of an actors saved feed preferences.
     /// </summary>
-    public IReadOnlyList<SavedFeedPreference2> Items { get; init; }
+    [JsonRequired]
+    public IReadOnlyList<SavedFeedPreferenceV2> Items { get; init; }
 }
 
 /// <summary>
 /// Encapsulates feed preferences for an actor
 /// </summary>
-public record SavedFeedPreference2 : Preference
+public sealed record SavedFeedPreferenceV2
 {
     /// <summary>
-    /// Creates a new instance of <see cref="SavedFeedPreference2"/>
+    /// Creates a new instance of <see cref="SavedFeedPreferenceV2"/>
     /// </summary>
     /// <param name="id">The identifier of the feed preference.</param>
     /// <param name="type">The type of the feed preference.</param>
     /// <param name="value">The value of the feed preference.</param>
     /// <param name="pinned">A flag indicating whether the feed is pinned.</param>
     [JsonConstructor]
-    public SavedFeedPreference2(string id, SavedFeedPreferenceType type, string value, bool pinned)
+    public SavedFeedPreferenceV2(string id, SavedFeedPreferenceType type, string value, bool pinned)
     {
         Id = id;
         Type = type;
@@ -57,21 +58,25 @@ public record SavedFeedPreference2 : Preference
     /// <summary>
     /// Gets the identifier of the feed preference.
     /// </summary>
+    [JsonRequired]
     public string Id { get; init; }
 
     /// <summary>
     /// Gets the type of the feed preference.
     /// </summary>
+    [JsonRequired]
     public SavedFeedPreferenceType Type { get; init; }
 
     /// <summary>
     /// Gets the value of the feed preference.
     /// </summary>
+    [JsonRequired]
     public string Value { get; init; }
 
     /// <summary>
     /// Gets a flag indicating whether the feed is pinned.
     /// </summary>
+    [JsonRequired]
     public bool Pinned { get; init; }
 }